Fixes:

* when url ends with slash first param route is the match (fix #1804)
* router should check if node is suitable as matching route by path+method and if not then continue search in tree (fix #1808)