@aldas 2021-03-16
benchmarks for
```bash
go test -run="-" -bench="BenchmarkEcho.*" -count 10
go test -run="-" -bench="BenchmarkRouter.*" -count 10
```

Results:
```bash
x@x:~/code/echo$ benchstat before_fix.out after_fix.out 
name                      old time/op    new time/op    delta
EchoStaticRoutes-6          17.0µs ± 2%    18.4µs ± 1%  +7.69%  (p=0.000 n=9+10)
EchoStaticRoutesMisses-6    17.3µs ± 3%    18.4µs ± 1%  +6.47%  (p=0.000 n=10+10)
EchoGitHubAPI-6             32.5µs ± 3%    34.0µs ± 1%  +4.58%  (p=0.000 n=8+9)
EchoGitHubAPIMisses-6       32.3µs ± 3%    34.3µs ± 1%  +6.46%  (p=0.000 n=10+10)
EchoParseAPI-6              2.04µs ± 2%    2.10µs ± 3%  +2.90%  (p=0.000 n=10+10)
 
name                         old time/op    new time/op    delta
RouterStaticRoutes-6           13.6µs ± 2%    14.8µs ± 1%   +8.51%  (p=0.000 n=10+10)
RouterStaticRoutesMisses-6      525ns ± 2%     586ns ± 3%  +11.71%  (p=0.000 n=10+10)
RouterGitHubAPI-6              25.9µs ± 2%    28.2µs ± 1%   +9.05%  (p=0.000 n=10+10)
RouterGitHubAPIMisses-6         616ns ± 1%     704ns ± 1%  +14.38%  (p=0.000 n=10+8)
RouterParseAPI-6               1.37µs ± 1%    1.46µs ± 2%   +6.86%  (p=0.000 n=10+10)
RouterParseAPIMisses-6          358ns ± 2%     361ns ± 1%   +0.87%  (p=0.004 n=9+10)
RouterGooglePlusAPI-6           882ns ± 2%     955ns ± 3%   +8.35%  (p=0.000 n=10+10)
RouterGooglePlusAPIMisses-6     493ns ± 0%     546ns ± 1%  +10.76%  (p=0.000 n=7+10)
RouterParamsAndAnyAPI-6        2.07µs ± 1%    2.35µs ± 0%  +13.58%  (p=0.000 n=10+8)
```