@lammel 2021-03-17
Sounds like we already have a plan on how paths are resolved for specific HTTP methods.
I need to check on the tests first.
This is definitely something the must be documented in the release notes and docs.

What are the opinions on changing this behaviour for v4.3 vs v5?