@aldas 2021-03-17
> Sounds like we already have a plan on how paths are resolved for specific HTTP methods.
> I need to check on the tests first.
> This is definitely something the must be documented in the release notes and docs.
> 
> What are the opinions on changing this behaviour for v4.3 vs v5?

nah, I am not that sure about it. I just wanted to have a note there. Thing is that Echo has gone so far without having to have more extensive error handling for app "setup" therefore question is if it is necessary at all?  At the same time it could be that we have pushed that burden to developer to debug their code and see why some things do not work.

For route building problems I see 3 options:
1) return error immediately (breaking change)
2) store error (like binder) and return on startup (there we are already returning errors) (probably breaking)
3) continue as today. maybe improve by logging errors/warnings

For example if you look what Gin or Chi does - none of them actually return error when adding a route.