@lammel 2021-03-17
For that we could split it up into 2 stops.

Provide route resolving based on method but
1. for v4.3 silently ignore errors on route building
2. for v5 change adding route to return errors

The changed behaviour is a correction in route handlin  (think bugfix) and can be landed in v4.3 with an important note IMHO.