@lammel 2021-04-25
Changes look good. Haven't done any debug testing, but tests are there and document the changed behavior pretty well.

We have some "breaking" changes actually:
- Correctly match routes with trailing backslash
- Correctly match routes with param node as leaf
- Routes with param node as leaf will capture remaining path into the last param node
- Considering handlers when resolving routes 
 
While the first is a correction and second can be considered fixes to existing bugs in routing, the 3rd point is change in behavior.
As the current misbehavior will not find a matching route, I think it is enough to add that to the release notes.

Did you do any performance tests for the changes? Benchstat tells us of a small drop for some scenarios.