@aldas 2021-04-27
I posted performance stats https://github.com/labstack/echo/pull/1812#issuecomment-800635619 up in this thread. It is below 10% increase but it is expected as previously method match was checked once at the end and now is checked when path is a match and does backtracking when it is but method for that matched route does not exist. This is inherently more work.

I would not say to these are huge breaking changes. more like corner cases.  Also `Routes with param node as leaf will capture remaining path into the last param node` is not new feature - this is how current version acts (on some cases).

I'll add those points as note about routing in changelog